# sam: optimize ParseAux