update #89

change on (newly added) aux parsing benchmark:

```
name        old time/op    new time/op    delta
ParseAux-4     661ns ± 3%     256ns ± 1%  -61.20%  (p=0.000 n=8+10)

name        old alloc/op   new alloc/op   delta
ParseAux-4      376B ± 0%      128B ± 0%  -65.96%  (p=0.000 n=10+10)

name        old allocs/op  new allocs/op  delta
ParseAux-4      12.0 ± 0%       7.0 ± 0%  -41.67%  (p=0.000 n=10+10)
```