@brentp 2017-04-11
with new benchmark this looks like:

```
name        old time/op    new time/op    delta
ParseAux-4    1.65µs ± 0%    1.08µs ± 5%  -34.42%  (p=0.000 n=9+10)

name        old alloc/op   new alloc/op   delta
ParseAux-4      816B ± 0%      416B ± 0%  -49.02%  (p=0.000 n=10+10)

name        old allocs/op  new allocs/op  delta
ParseAux-4      26.0 ± 0%      19.0 ± 0%  -26.92%  (p=0.000 n=10+10)
```