@brentp 2017-04-11
with benchmarks split out:
```
name             old time/op    new time/op    delta
ParseAuxInt-4       200ns ± 4%      81ns ± 0%  -59.57%  (p=0.000 n=10+10)
ParseAuxZ-4         220ns ± 1%      84ns ± 1%  -61.77%  (p=0.000 n=10+9)
ParseAuxFloat-4     270ns ± 1%     156ns ± 1%  -42.31%  (p=0.000 n=10+10)
ParseAuxArray-4     670ns ± 1%     548ns ± 0%  -18.11%  (p=0.000 n=9+8)

name             old alloc/op   new alloc/op   delta
ParseAuxInt-4       96.0B ± 0%     16.0B ± 0%  -83.33%  (p=0.000 n=10+10)
ParseAuxZ-4          147B ± 0%       64B ± 0%  -56.46%  (p=0.000 n=10+10)
ParseAuxFloat-4      104B ± 0%       24B ± 0%  -76.92%  (p=0.000 n=10+10)
ParseAuxArray-4      336B ± 0%      256B ± 0%  -23.81%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
ParseAuxInt-4        4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=10+10)
ParseAuxZ-4          4.00 ± 0%      2.00 ± 0%  -50.00%  (p=0.000 n=10+10)
ParseAuxFloat-4      4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=10+10)
ParseAuxArray-4      10.0 ± 0%       9.0 ± 0%  -10.00%  (p=0.000 n=10+10)
```